[refactor] change name from http_sd to registry
