+++ hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/registry/discovery/DiscoveryClientManagement.java
@@ -17,39 +17,39 @@
  * under the License.
  */
 
-package org.apache.hertzbeat.collector.collect.httpsd.discovery;
+package org.apache.hertzbeat.collector.collect.registry.discovery;
 
 import java.util.Objects;
-import org.apache.hertzbeat.collector.collect.httpsd.constant.DiscoveryClientInstance;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.impl.ConsulDiscoveryClient;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.impl.NacosDiscoveryClient;
-import org.apache.hertzbeat.common.entity.job.protocol.HttpsdProtocol;
+import org.apache.hertzbeat.collector.collect.registry.constant.DiscoveryClientInstance;
+import org.apache.hertzbeat.collector.collect.registry.discovery.impl.ConsulDiscoveryClient;
+import org.apache.hertzbeat.collector.collect.registry.discovery.impl.NacosDiscoveryClient;
+import org.apache.hertzbeat.common.entity.job.protocol.RegistryProtocol;
 
 /**
  * Discovery Client Management
  */
 public class DiscoveryClientManagement {
 
-    public DiscoveryClient getClient(HttpsdProtocol httpsdProtocol) {
-        return createClient(httpsdProtocol, DiscoveryClientInstance.getByName(httpsdProtocol.getDiscoveryClientTypeName()));
+    public DiscoveryClient getClient(RegistryProtocol registryProtocol) {
+        return createClient(registryProtocol, DiscoveryClientInstance.getByName(registryProtocol.getDiscoveryClientTypeName()));
     }
 
-    private DiscoveryClient createClient(HttpsdProtocol httpsdProtocol, DiscoveryClientInstance discoveryClientInstance) {
+    private DiscoveryClient createClient(RegistryProtocol registryProtocol, DiscoveryClientInstance discoveryClientInstance) {
         if (Objects.equals(discoveryClientInstance, DiscoveryClientInstance.NOT_SUPPORT)) {
             return null;
         }
 
-        return doCreateClient(httpsdProtocol, discoveryClientInstance);
+        return doCreateClient(registryProtocol, discoveryClientInstance);
     }
 
-    private DiscoveryClient doCreateClient(HttpsdProtocol httpsdProtocol, DiscoveryClientInstance discoveryClientInstance) {
+    private DiscoveryClient doCreateClient(RegistryProtocol registryProtocol, DiscoveryClientInstance discoveryClientInstance) {
         DiscoveryClient discoveryClient;
         switch (discoveryClientInstance) {
             case CONSUL -> discoveryClient = new ConsulDiscoveryClient();
             case NACOS -> discoveryClient = new NacosDiscoveryClient();
             default -> { return null; }
         }
-        discoveryClient.initClient(discoveryClient.buildConnectConfig(httpsdProtocol));
+        discoveryClient.initClient(discoveryClient.buildConnectConfig(registryProtocol));
         return discoveryClient;
     }
 }
