+++ hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/registry/discovery/entity/ConnectConfig.java
@@ -17,21 +17,19 @@
  * under the License.
  */
 
-package org.apache.hertzbeat.collector.collect.httpsd.constant;
+package org.apache.hertzbeat.collector.collect.registry.discovery.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
 
 /**
- * Discovery client instance status.
+ * Discovery Client Connect Config
  */
-public final class DiscoveryClientHealthStatus {
-
-    /**
-	 * Discovery client instance status is UP.
-	 */
-    public static final String UP = "UP";
-
-    /**
-	 * Discovery client instance status is DOWN.
-	 */
-    public static final String DOWN = "DOWN";
-
+@Data
+@Builder
+@AllArgsConstructor
+public class ConnectConfig {
+    private String host;
+    private int port;
 }
