+++ hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/registry/discovery/impl/NacosDiscoveryClient.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hertzbeat.collector.collect.httpsd.discovery.impl;
+package org.apache.hertzbeat.collector.collect.registry.discovery.impl;
 
 import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.api.naming.NamingFactory;
@@ -27,12 +27,12 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.hertzbeat.collector.collect.httpsd.constant.DiscoveryClientHealthStatus;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.DiscoveryClient;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.entity.ConnectConfig;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.entity.ServerInfo;
-import org.apache.hertzbeat.collector.collect.httpsd.discovery.entity.ServiceInstance;
-import org.apache.hertzbeat.common.entity.job.protocol.HttpsdProtocol;
+import org.apache.hertzbeat.collector.collect.registry.constant.DiscoveryClientHealthStatus;
+import org.apache.hertzbeat.collector.collect.registry.discovery.DiscoveryClient;
+import org.apache.hertzbeat.collector.collect.registry.discovery.entity.ConnectConfig;
+import org.apache.hertzbeat.collector.collect.registry.discovery.entity.ServerInfo;
+import org.apache.hertzbeat.collector.collect.registry.discovery.entity.ServiceInstance;
+import org.apache.hertzbeat.common.entity.job.protocol.RegistryProtocol;
 
 /**
  * DiscoveryClient impl of Nacos
@@ -43,10 +43,10 @@ public class NacosDiscoveryClient implements DiscoveryClient {
     private ConnectConfig localConnectConfig;
 
     @Override
-    public ConnectConfig buildConnectConfig(HttpsdProtocol httpsdProtocol) {
+    public ConnectConfig buildConnectConfig(RegistryProtocol registryProtocol) {
         return ConnectConfig.builder()
-                .host(httpsdProtocol.getHost())
-                .port(Integer.parseInt(httpsdProtocol.getPort()))
+                .host(registryProtocol.getHost())
+                .port(Integer.parseInt(registryProtocol.getPort()))
                 .build();
     }
