+++ hertzbeat-collector/hertzbeat-collector-common/src/main/java/org/apache/hertzbeat/collector/dispatch/DispatchConstants.java
@@ -120,9 +120,9 @@ public interface DispatchConstants {
      */
     String PROTOCOL_POP3 = "pop3";
     /**
-     * protocol http_sd
+     * protocol registry
      */
-    String PROTOCOL_HTTP_SD = "httpsd";
+    String PROTOCOL_REGISTRY = "registry";
     /**
      * protocol redfish
      */
